Cbrdx th adm robertochajin

+++ Dominio/src/main/java/com/ciberdix/th/repositories/refactor/VTercerosInmueblesRefactorRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 /**
  * Created by robertochajin on 4/04/17.
  */
-public interface VTercerosInmueblesRefactorRepository extends CrudRepository<VTercerosInmuebles, Long> {
+public interface VTercerosInmueblesRefactorRepository extends CrudRepository<VTercerosInmuebles, Integer> {
     List<VTercerosInmuebles> findByIdTercero(Long id);
 }
